% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpg_table.R
\name{browse_table}
\alias{browse_table}
\title{Browse table returned by \code{\link[=cjpg_table]{cjpg_table()}} or \code{\link[=cjsg_table]{cjsg_table()}}}
\usage{
browse_table(table, patterns)
}
\arguments{
\item{table}{Table returned by \code{\link[=cjpg_table]{cjpg_table()}} or \code{\link[=cjsg_table]{cjsg_table()}}
(only valid for \code{"classes"} or \code{"subjects"} types)}

\item{patterns}{A list containing (at most) 6 character vectors
of one or more regular expressions (applied from left to right
on \code{name0} to \code{name5}), e.g.,
\code{list(c("ADM", "CRIMINAL"), "", "", "", "", "Recurso")}}
}
\value{
The original table filtered according to \code{patterns}
}
\description{
This function uses a list of regex to filter CJPG
and CJSG tables (only if they are of type \code{"classes"} or
\code{"subjects"}) more easily than with \code{dplyr::select()}. For
details on how the matching occurs, see \strong{Matching}.
}
\section{Matching}{
 For the matching to work properly, \code{patterns}
should be a list of at most 6 character vectors, each one
containing either one or a vector of regular expressions to
be applied from left to right on columns \code{name0} to \code{name5}
(note that vectors are ORed and different elements are ANDed).
Example: If \code{patterns} looks something like
\code{list(c("ADM", "CRIMINAL"), "", "", "", "", "Recurso")},
then we'll get back the rows where \code{name0} contains "ADM"
\strong{or} "CRIMINAL" \strong{and} where \code{name5} contains "Recurso".
}

\seealso{
\code{\link[=cjpg_table]{cjpg_table()}}, \code{\link[=cjsg_table]{cjsg_table()}}
}
