% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cposg.R
\name{build_url_cpo_sg}
\alias{build_url_cpo_sg}
\title{Funcao que faz o download das informacoes de um processo de segundo
grau (SG) no TJSP.}
\usage{
build_url_cpo_sg(p, tj)
}
\description{
Funcao que faz o download das informacoes de um processo de segundo
grau (SG) no TJSP.
}
