% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpg.R
\name{cjpg}
\alias{cjpg}
\title{Consulta de Julgados de Primeiro Grau}
\usage{
cjpg(session, parms, path = "data-raw/cjpg", max_pag = 10,
  overwrite = FALSE, verbose = TRUE, p = 0.05)
}
\arguments{
\item{session}{sess??o da CJPG, retornado pela fun????o \code{\link{cjpg_session}}.}

\item{parms}{form com par??metros de busca, retornado pela fun????o \code{\link{cjpg_parms}}.}

\item{path}{caminho onde os arquivos HTML ser??o baixados. Tentar?? criar a pasta automaticamente.}

\item{max_pag}{n??mero m??ximo de p??ginas a serem baixadas. Se \code{NA} ou \code{Inf}, baixar?? todas as p??ginas resultantes da pesquisa.}

\item{overwrite}{logical: sobrescrever arquivos salvos?}

\item{verbose}{logical: imprimir mensagens de acompanhamento do download?}

\item{p}{probabilidade de imprimir uma mensagem de acompanhamento se \code{verbose} for \code{TRUE}. Default 5\%.}
}
\description{
Baixa arquivos HTML correspondentes a uma busca na CJPG.
}
