% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpg-tjsp.R
\name{cjpg_parms}
\alias{cjpg_parms}
\title{Par??metros de pesquisa}
\usage{
cjpg_parms(session, livre = "", data_inicial = NULL, data_final = NULL,
  secoes = "", classes = "", assuntos = "", magistrados = "",
  varas = "")
}
\arguments{
\item{session}{sess??o da CJPG.}

\item{livre}{string com pesquisa livre por palavras-chave.}

\item{data_inicial}{data inicial da decis??o em formado \%Y-\%m-\%d ou objeto Date.}

\item{data_final}{data final da decis??o em formado \%Y-\%m-\%d ou objeto Date.}

\item{secoes}{sess??es}

\item{classes}{vetor com as classes que se deseja pesquisar. Obter o vetor de \code{cjpg_classes}.}

\item{assuntos}{vetor com os assuntos que se deseja pesquisar. Obter o vetor de \code{cjpg_assuntos}.}

\item{magistrados}{vetor com os magistrados que se deseja pesquisar. NOT IMPLEMENTED YET.}

\item{varas}{vetor com as varas que se deseja pesquisar. Obter o vetor de \code{cjpg_varas}.}
}
\value{
objeto de classe \code{form}.
}
\description{
Inclui par??metros de pesquisa da CJPG.
}
\details{
O intervalo de datas, caso seja inclu??do, precisa ser de no m??ximo um ano.
}
\examples{
\dontrun{
library(tjsp)
s <- cjpg_session()
parms <- cjsg_parms(
  s, livre = 'sentenca', data_inicial = '2015-01-01',
  data_final = '2015-05-01')
parms
cjpg(s, parms, max_pag = 1, path = '.')
}
}
