% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpg_table.R
\name{cjpg_table}
\alias{cjpg_table}
\title{Download information about some of CJPG's structures}
\usage{
cjpg_table(type, tj = "tjsp")
}
\arguments{
\item{type}{Either \code{"classes"} or \code{"subjects"} or \code{"courts"}}

\item{tj}{TJ from which to get data (only works with TJSP for now)}
}
\value{
A tibble with either 12 columns (if \code{type} is \code{"classes"}
or \code{"subjects"}) or 3 columns (if \code{type} is \code{"courts"})
}
\description{
Downloads a table with information about lawsuit's
classes, subjects or courts to help with \code{\link[=download_cjpg]{download_cjpg()}}. You
can also browse some of these tables with \code{\link[=browse_table]{browse_table()}}.
}
\seealso{
\code{\link[=download_cjpg]{download_cjpg()}}, \code{\link[=browse_table]{browse_table()}}
}
