% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsg-tjsp.R
\name{cjsg_npags}
\alias{cjsg_npags}
\title{N??mero de p??ginas}
\usage{
cjsg_npags(session, parms = NULL)
}
\arguments{
\item{session}{sess??o da CJSG.}

\item{parms}{se for \code{NULL}, admite que j?? est?? na p??gina de resultados da CJSG.}
}
\value{
objeto de classe \code{session}.
}
\description{
Calcula o n??mero de p??ginas retornadas por uma consulta de julgados de segundo grau.
}
\examples{
\dontrun{
library(tjsp)
s <- cjsg_session()
parms <- cjsg_parms(s, livre = 'acordam')
cjsg_npags(s, parms)
cjsg(s, parms, max_pag = 1, path = '.') # chamada internamente dentro de cjsg.
}

}
