% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsg-tjsp.R
\name{cjsg_pag}
\alias{cjsg_pag}
\title{Baixa uma p??gina.}
\usage{
cjsg_pag(pag, path, ow, s)
}
\arguments{
\item{pag}{n??mero da p??gina a ser baixada.}

\item{path}{pasta em que o arquivo ser?? salvo.}

\item{ow}{logical, sobrescrever arquivo?}

\item{s}{sessao da CJSG.}
}
\description{
Baixa uma p??gina a partir de um resultado da CJSG. N??o deve ser usado diretamente.
}
