% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects-2inst.R
\name{cjsg_tables}
\alias{cjsg_tables}
\title{Downloads table of CJSG items}
\usage{
cjsg_tables(tipo = c("classes", "assuntos"), tj = "tjsp")
}
\arguments{
\item{tipo}{class, subject or circuit}

\item{tj}{TJ of the table (only works with TJSP for now)}
}
\value{
\code{data.frame} with titles and ID of the leaves and nodes of the table tree
}
\description{
Downloads a table with the elements of class, subject and circuit to help with \code{\link[=cjsg]{cjsg()}}
}
