% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects-2inst-tjsp.R
\name{cjsg_tabs}
\alias{cjsg_tabs}
\title{Baixa tabela de itens da CJPG.}
\usage{
cjsg_tabs(tipo = c("classes", "assuntos"))
}
\arguments{
\item{tipo}{classe, assunto ou varas}
}
\value{
\code{data.frame} com titulos e cods das folhas e titulos e cods dos pais (??ndice 0 ?? a raiz e quanto menor o ??ndice, mais pr??ximo da raiz)
}
\description{
Baixa uma tabela com os elementos de classe, assunto ou varas, para ajudar a baixar usando a fun????o CJPG.
}
