% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsg_aux.R
\name{courts_table}
\alias{courts_table}
\title{Download table with court information for \code{\link[=cjsg_table]{cjsg_table()}}}
\usage{
courts_table()
}
\description{
Download table with court information for \code{\link[=cjsg_table]{cjsg_table()}}
}
