% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cposg.R
\name{cpo_sg}
\alias{cpo_sg}
\title{Funcao que faz o download das informacoes de um processo de segundo
grau (SG) no TJSP.}
\usage{
cpo_sg(processos, path = "data-raw/cpo-sg", tj = "TJSP")
}
\description{
Retorna um data.frame com os metadados basicos e andamentos do processo
}

