% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cposg-tjsp.R
\name{cposg}
\alias{cposg}
\title{Baixa processos}
\usage{
cposg(processos, path = "data-raw/cposg", overwrite = FALSE)
}
\arguments{
\item{processos}{n??mero do processo (string apenas com os n??meros).}

\item{path}{caminho da pasta onde os arquivos HTML ser??o salvos.}

\item{overwrite}{sobrescrever os arquivos HTML?}
}
\description{
Baixa processos na consulta de processos de segundo grau do TJSP.
}
