% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cposg-tjsp.R
\name{cposg_um}
\alias{cposg_um}
\title{Baixa um processo.}
\usage{
cposg_um(p, path, ow)
}
\arguments{
\item{p}{n??mero do processo (string apenas com os n??meros).}

\item{path}{caminho da pasta onde ser?? salvo o arquivo HTML.}

\item{ow}{sobrescrever o arquivo HTML?}
}
\description{
Baixa um processo na consulta de processos de segundo grau do TJSP. Deve ser usado internamente.
}
