% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dje.R
\name{dje}
\alias{dje}
\alias{dje_range}
\title{Baixa diários oficiais}
\usage{
dje(tj = "TJSP", dates = Sys.Date(), path = "data-raw/dje_pdf",
  verbose = FALSE)

dje_range(from, to, tj = "TJSP", path = "data-raw/dje_pdf",
  verbose = FALSE)
}
\arguments{
\item{tj}{character vector indicando o Tribunal. Atualmente funciona com
TJSP, TJAC, TJAL, TJAM, TJMS, TJRN, TJSC, TJCE, TJBA. Default \code{'TJSP'}.}

\item{dates}{Date vector ou character vector em YYYY-MM-DD com as datas
que se deseja baixar. Default \code{Sys.Date()}.}

\item{path}{pasta onde os arquivos serão gravados. Para cada data, uma pasta
será criada e os arquivos PDF serão salvos nessa pasta. Default \code{'data-raw/dje_pdf'}}

\item{verbose}{imprimir mensagens? Default \code{FALSE}.}

\item{from, to}{Date vector ou character vector em formato YYYY-MM-DD.}
}
\value{
\code{tbl_df} com diagnóstico dos resultados.
}
\description{
Acessa os Diários de Justiça Eletrônicos dos Tribunais de Justiça e baixa
os arquivos em PDF.
}
\examples{

dir.create('data-raw/dje_pdf', recursive = TRUE, showWarnings = FALSE)
tjsp_dje <- dje(dates = Sys.Date() - 0:3)
table(tjsp_dje$result)

# --------------------------------------------------------------------------
tjal_dje <- dje(tj = 'TJAL', dates = Sys.Date() - 0:3)
tjam_dje <- dje(tj = 'TJAM', dates = Sys.Date() - 0:3)
tjce_dje <- dje(tj = 'TJCE', dates = Sys.Date() - 0:3)
tjba_dje <- dje(tj = 'TJBA', dates = Sys.Date() - 0:3)
tjmt_dje <- dje(tj = 'TJMT', dates = Sys.Date() - 0:3)
tjrn_dje <- dje(tj = 'TJRN', dates = Sys.Date() - 0:3)

}

