% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpg.R
\name{download_cjpg}
\alias{download_cjpg}
\title{Dowload CJPG search results}
\usage{
download_cjpg(query, path = ".", classes = "", subjects = "",
  courts = "", date_start = "", date_end = "", min_page = 1,
  max_page = 1, cores = 1, tj = "tjsp")
}
\arguments{
\item{query}{Character vector wtih search query}

\item{path}{Path to directory where to save HTMLs}

\item{classes}{Character vector with lawsuit class IDs}

\item{subjects}{Character vector with lawsuit subject IDs}

\item{courts}{Character vector with lawsuit court IDs}

\item{date_start}{Lower bound for date}

\item{date_end}{Upper bound for date}

\item{min_page}{First page of results to download}

\item{max_page}{Last page of results to download}

\item{cores}{The number of cores to be used when downloading}

\item{tj}{TJ from which to get data (only works with TJSP for now)}
}
\description{
Dowload CJPG search results
}
\seealso{
\code{\link[=cjpg_table]{cjpg_table()}}
}
