% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsg.R
\name{download_cjsg}
\alias{download_cjsg}
\title{Dowload CJSG search results}
\usage{
download_cjsg(query = "", path, classes = "", subjects = "",
  courts = "", trial_start = "", trial_end = "",
  registration_start = "", registration_end = "", min_page = 1,
  max_page = 1, cores = 1, tj = "tjsp")
}
\arguments{
\item{query}{Character vector wtih search query}

\item{path}{Path to directory where to save HTMLs}

\item{classes}{Character vector with lawsuit class IDs}

\item{subjects}{Character vector with lawsuit subject IDs}

\item{courts}{Character vector with lawsuit court IDs}

\item{trial_start}{Date when the trial started}

\item{trial_end}{Date when the trial ended}

\item{registration_start}{Date when registration started}

\item{registration_end}{Date then registration ended}

\item{min_page}{First page of results to download}

\item{max_page}{Last page of results to download}

\item{cores}{The number of cores to be used when downloading}

\item{tj}{TJ form which to get data (only works with TJSP for now)}
}
\description{
Dowload CJSG search results
}
\seealso{
\code{\link[=cjsg_table]{cjsg_table()}}
}
