% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsg.R
\name{download_cjsg_tjmg}
\alias{download_cjsg_tjmg}
\title{Temporary function for downloading TJMG's CJSG queries}
\usage{
download_cjsg_tjmg(query, path = ".", classes = "", subjects = "",
  courts = "", trial_start = "", trial_end = "",
  registration_start = "", registration_end = "", min_page = 1,
  max_page = 1, rapporteurs = "")
}
\arguments{
\item{query}{Character vector with search query}

\item{path}{Path to directory where to save HTMLs}

\item{classes}{Character vector with class IDs (e.g. \code{c(175, 43, 259, 263)})}

\item{subjects}{Character vector with subject IDs (e.g. \code{c(10207, 10008, 10199)})}

\item{courts}{Character vector with court IDs (e.g. \code{c("1-7", "1-9", "2-3", "1-1")})}

\item{trial_start}{Lower bound for trial date}

\item{trial_end}{Upper bound for trial date}

\item{registration_start}{Lower bound for registration date}

\item{registration_end}{Upper bound for registration date}

\item{min_page}{First page of results to download}

\item{max_page}{Last page of results to download. If is \code{NA} or
\code{Inf}, we use \code{\link{peek_cjsg}}.}

\item{rapporteurs}{Character vector with rapporteur IDs (e.g. \code{c("2-1528561", "2-2345361")})}
}
\value{
A character vector with the paths to the downloaded files
}
\description{
Temporary function for downloading TJMG's CJSG queries
}
