% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{download_decision}
\alias{download_decision}
\title{Download decision PDFs}
\usage{
download_decision(id, path, tj = "tjsp")
}
\arguments{
\item{id}{A character vector with decision IDs}

\item{path}{Path to directory where to save PDFs}

\item{tj}{TJ to download decisions (only works with TJSP for now)}
}
\description{
Download decision PDFs
}
