% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{download_decisions}
\alias{download_decisions}
\title{Download decisions}
\usage{
download_decisions(cod_decision, tj = "tjsp", path = ".")
}
\arguments{
\item{cod_decision}{Column cd_acordao of d_cjsg}

\item{tj}{TJ to download decisions (only works with TJSP for now)}

\item{path}{Path to the directory where the lawsuit should be downloaded}
}
\description{
Download decisions from TJSP
}
