% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision-tjsp.R
\name{download_decisions_tjsp}
\alias{download_decisions_tjsp}
\title{Download decisions}
\usage{
download_decisions_tjsp(cod_decision, path = ".")
}
\arguments{
\item{cod_decision}{column cd_acordao of d_cjsg}

\item{path}{folder}
}
\description{
Download decisions from TJSP
}
