% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dje.R
\name{download_dje}
\alias{download_dje}
\title{Download Diarios da Justica Eletronicos (DJEs) from a Tribunal
de Justica (TJ)}
\usage{
download_dje(tj, dates = Sys.Date(), path = ".", verbose = FALSE)
}
\arguments{
\item{tj}{Abbreviation of a Justice Court (i.e. "TJSP")}

\item{dates}{A character vector of dates of the form \emph{YYYY-MM-DD}}

\item{path}{Path to the directory where the DJEs should be downloaded}

\item{verbose}{Whether to print download statuses}
}
\value{
A character vector with the paths to the downloaded DJEs
}
\description{
This function downloads Diarios da Justica Eletronicos
(Electronic Court Records) as PDFs. Given a Tribunal de Justica
(Justice Court), a character vector of dates, and the path to a
directory it will collect the DJEs published on these dates, and
save them to the provided directory.
}
\section{About DJEs}{

Electronic Court Records are documents published almost
daily by every state's Justice Court containing important information
regarding new lawsuits being filed, decisions being made by judges,
and so on. Note that DJEs may be split into multiple "booklets"; when
this happens it means that each booklet contains information about
only one type of lawsuit.
}

\section{Implemented TJs}{

Unfortunatelly \code{download_dje()} doesn't yet work with all 27 TJs in
Brazil. Here are the ones already implemented:
\itemize{
\item TJAC: Acre
\item TJAL: Alagoas
\item TJAM: Amazonas
\item TJBA: Bahia
\item TJCE: Ceara
\item TJMS: Mato Grosso do Sul
\item TJRN: Rio Grande do Norte
\item TJSC: Santa Catarina
\item TJSP: Sao Paulo
}
}

\examples{
\dontrun{
download_dje("TJSP", Sys.Date())
download_dje("TJSC", Sys.Date() - 0:3, verbose = TRUE)
}

}
