% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteiro-teor-tjsp.R
\name{download_pdf}
\alias{download_pdf}
\title{Download PDF of a decision}
\usage{
download_pdf(codes, dest = ".", giveup = 10, wait = 5,
  overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{codes}{character vector containing id of the cases, obtainable from
\code{\link{cjsg}} function.}

\item{dest}{destination folder of the pdf file}

\item{giveup}{Give up after \code{giveup} failed transfers}

\item{wait}{seconds to wait between failed transfers}

\item{overwrite}{overwrite?}

\item{verbose}{print error messages}
}
\description{
Download PDF of a decision (second instance)
}
