% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects-2inst-tjsp.R
\name{list_secoes_2inst}
\alias{list_secoes_2inst}
\title{Se????es de segunda inst??ncia.}
\usage{
list_secoes_2inst()
}
\value{
\code{tibble} com as colunas \code{pai}, \code{secao} e \code{cod}. Usualmente \code{cod} ser?? usado na fun????o \code{\link{cjsg_parms}}
}
\description{
Baixa lista das se????es de segunda inst??ncia.
}
