% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{make_parser}
\alias{make_parser}
\title{Create a parser skeleton}
\usage{
make_parser(type = "cposg")
}
\arguments{
\item{type}{The type of parser (\code{"cpopg"} or \code{"cposg"})}
}
\description{
Create a parser skeleton
}
