% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-cjpg.R
\name{parse_cjpg}
\alias{parse_cjpg}
\title{CJPG parser}
\usage{
parse_cjpg(arqs)
}
\arguments{
\item{arqs}{Character vector with the paths of the files to be parsed}
}
\value{
A tibble with the columns
\itemize{
\item \code{arq} Name of the file
\item \code{id} ID found in the page read
\item \code{cd_acordao} Unique ID of the ruling
\item \code{n_processo} Number of the lawsuit (doesn't have to be unique)
\item \code{comarca} Name of the district
\item \code{data_julgamento} Date of the judgement (\%d/\%m/\%Y)
\item \code{data_registro} Date of registration in the system (\%d/\%m/\%Y)
\item \code{ementa} Summary of the ruling
\item \code{orgao_julgador} Body responsible for the appeal
\item \code{outros_numeros} Old/additional IDs
\item \code{relatora} Name of the rapporteur
\item \code{classe_assunto} Class/subject, separated by slashes
\item \code{txt_ementa} Text of the summary with no formatting
}
}
\description{
Parser for files downloaded by \code{\link[=cjpg]{cjpg()}}.
}
