% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cjpg.R
\name{parse_cjpg_}
\alias{parse_cjpg_}
\title{Parse a page of CJPG results}
\usage{
parse_cjpg_(file, pb = NULL)
}
\arguments{
\item{file}{The path to the file to be parsed}

\item{pb}{Progress bar created by \code{\link[=parse_cjpg]{parse_cjpg()}}}
}
\value{
A tibble with the parsed information
}
\description{
Parse a page of CJPG results
}
