% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cjpg.R
\name{parse_cjpg_lawsuit}
\alias{parse_cjpg_lawsuit}
\title{Parse one lawsuit from a CJSG page}
\usage{
parse_cjpg_lawsuit(node)
}
\arguments{
\item{node}{A \code{.fundocinza1} node extracted from the page}
}
\value{
One row with the data concerning the lawsuit
}
\description{
Parse one lawsuit from a CJSG page
}
