% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cjsg.R
\name{parse_cjsg}
\alias{parse_cjsg}
\title{Parse lawsuits extracted from CJSG query}
\usage{
parse_cjsg(file, cores = 1)
}
\arguments{
\item{file}{Character vector with the paths to one or more files}

\item{cores}{Number of cores to use when parsing}
}
\value{
A tibble with the columns
\itemize{
\item \code{file} Name of the file
\item \code{id_page} ID found in the page
\item \code{id_decision} Unique ID of the ruling
\item \code{id_lawsuit} Number of the lawsuit (doesn't have to be unique)
\item \code{class_subject} Class/subject, separated by slashes
\item \code{district} Name of the district
\item \code{court} Body responsible for the appeal
\item \code{date_decision} Date of the judgement (\%d/\%m/\%Y)
\item \code{date_publication} Date of the publication (\%d/\%m/\%Y)
\item \code{date_registration} Date of registration in the system (\%d/\%m/\%Y)
\item \code{rapporteur} Name of the rapporteur
\item \code{summary} Summary of the ruling
\item \code{txt_summary} Text of the summary with no formatting
}
}
\description{
Parse lawsuits extracted from CJSG query
}
