% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cjsg.R
\name{parse_cjsg_}
\alias{parse_cjsg_}
\title{Parse a page of CJSG results}
\usage{
parse_cjsg_(file, pb = NULL)
}
\arguments{
\item{file}{The path to the file to be parsed}

\item{pb}{Progress bar created by \code{\link[=parse_cjsg]{parse_cjsg()}}}
}
\value{
A tibble with the parsed information
}
\description{
Parse a page of CJSG results
}
