% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_cpopg_all}
\alias{parse_cpopg_all}
\title{Shortcut for creating and running a complete CPOPG parser}
\usage{
parse_cpopg_all(file, path = ".", cores = 1)
}
\arguments{
\item{file}{A character vector with the paths to one ore more files}

\item{path}{The path to a directory where to save RDSs}

\item{cores}{The number of cores to be used when parsing}
}
\description{
Shortcut for creating and running a complete CPOPG parser
}
