% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cpopg.R
\name{parse_data.cpopg}
\alias{parse_data.cpopg}
\title{Parse data}
\usage{
\method{parse_data}{cpopg}(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parse data
}
