% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cposg.R
\name{parse_data.cposg}
\alias{parse_data.cposg}
\title{Parses data}
\usage{
\method{parse_data}{cposg}(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parses data
}
