% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cposg.R
\name{parse_decisions.cposg}
\alias{parse_decisions.cposg}
\title{Parses decisions}
\usage{
\method{parse_decisions}{cposg}(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parses decisions
}
