% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_parser.R
\name{parse_decisions}
\alias{parse_decisions}
\title{Parse decisions}
\usage{
parse_decisions(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parse decisions
}
