% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cpopg.R
\name{parse_hearings.cpopg}
\alias{parse_hearings.cpopg}
\title{Parse hearings}
\usage{
\method{parse_hearings}{cpopg}(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parse hearings
}
