% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_parser.R
\name{parse_hearings}
\alias{parse_hearings}
\title{Parse hearings}
\usage{
parse_hearings(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parse hearings
}
