% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cpopg.R
\name{parse_hist.cpopg}
\alias{parse_hist.cpopg}
\title{Parse history}
\usage{
\method{parse_hist}{cpopg}(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parse history
}
