% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_parser.R
\name{parse_hist}
\alias{parse_hist}
\title{Parse history}
\usage{
parse_hist(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parse history
}
