% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cpopg.R
\name{parse_movs.cpopg}
\alias{parse_movs.cpopg}
\title{Parse movements}
\usage{
\method{parse_movs}{cpopg}(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parse movements
}
