% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cposg.R
\name{parse_movs.cposg}
\alias{parse_movs.cposg}
\title{Parses movements}
\usage{
\method{parse_movs}{cposg}(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parses movements
}
