% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-cposg.R
\name{parse_movs}
\alias{parse_movs}
\title{Parses movements}
\usage{
parse_movs(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parses movements
}
