% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cpopg.R
\name{parse_parts.cpopg}
\alias{parse_parts.cpopg}
\title{Parse parts}
\usage{
\method{parse_parts}{cpopg}(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parse parts
}
