% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cposg.R
\name{parse_parts.cposg}
\alias{parse_parts.cposg}
\title{Parses parts}
\usage{
\method{parse_parts}{cposg}(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parses parts
}
