% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-cposg.R
\name{parse_parts}
\alias{parse_parts}
\title{Parses parts}
\usage{
parse_parts(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parses parts
}
