% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cpopg.R
\name{parse_pd.cpopg}
\alias{parse_pd.cpopg}
\title{Parse police department}
\usage{
\method{parse_pd}{cpopg}(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parse police department
}
