% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_parser.R
\name{parse_pd}
\alias{parse_pd}
\title{Parse police department}
\usage{
parse_pd(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parse police department
}
