% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpg.R
\name{peek_cjpg}
\alias{peek_cjpg}
\title{Check certain characteristics regarding a CJPG download}
\usage{
peek_cjpg(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[=download_cjpg]{download_cjpg()}} (
\code{path} will be ignored)}
}
\description{
Check certain characteristics regarding a CJPG download
}
\seealso{
\code{\link[=download_cjpg]{download_cjpg()}}, \code{\link[=cjpg_table]{cjpg_table()}}
}
