% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsg.R
\name{peek_cjsg}
\alias{peek_cjsg}
\title{Check certain characteristics regarding a CJSG download}
\usage{
peek_cjsg(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[=download_cjsg]{download_cjsg()}} (
\code{path} will be ignored)}
}
\description{
Check certain characteristics regarding a CJSG download
}
\seealso{
\code{\link[=download_cjsg]{download_cjsg()}}, \code{\link[=cjsg_table]{cjsg_table()}}
}
