% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-cposg.R
\name{run_parser}
\alias{run_parser}
\title{Runs a parser}
\usage{
run_parser(files, parser, cores = 1)
}
\arguments{
\item{files}{A character vector with the paths to one ore more files}

\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}

\item{cores}{The number of cores to be used when parsing}
}
\description{
Runs a parser
}
