% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpg_table.R
\name{browse_table}
\alias{browse_table}
\title{Browse table returned by \code{\link[=cjpg_table]{cjpg_table()}} or \code{\link[=cjsg_table]{cjsg_table()}}}
\usage{
browse_table(table, patterns)
}
\arguments{
\item{table}{Table returned by \code{\link[=cjpg_table]{cjpg_table()}} or \code{\link[=cjsg_table]{cjsg_table()}}
(only valid for \code{"classes"} or \code{"subjects"} types)}

\item{patterns}{A list containing (at most) 6 character vectors
of one or more regular expressions (applied from left to right
on root to leaves); e.g.
\code{list(c("ADM", "CRIMINAL"), "", "", "", "", "")}}
}
\description{
Browse table returned by \code{\link[=cjpg_table]{cjpg_table()}} or \code{\link[=cjsg_table]{cjsg_table()}}
}
\details{
Regex of the same level will be ORed and of different
levels will be ANDed, e.g.,
\code{list(c("ADM", "CRIMINAL"), "", "", "", "", "")} becomes
\code{list("(?:ADM|CRIMINAL)", "", "", "", "", "")} and each
element will be applied with \code{\link[dplyr:filter]{dplyr::filter()}} to a level of
the tree
}
\seealso{
\code{\link[=cjpg_table]{cjpg_table()}}, \code{\link[=cjsg_table]{cjsg_table()}}
}
