% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects-1inst-tjsp.R
\name{cjpg_table}
\alias{cjpg_table}
\title{Downloads table of CJPG items}
\usage{
cjpg_table(tipo = c("classe", "assunto", "varas"), tj = "tjsp")
}
\arguments{
\item{tipo}{class, subject or circuit}

\item{tj}{TJ of the table (only works with TJSP for now)}
}
\value{
\code{data.frame} with titles and ID of the leaves and nodes of the table tree
}
\description{
Downloads a table with the elements of class, subject and circuit to help with \code{\link[=cjpg]{cjpg()}}
}
