% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpg_table.R
\name{cjpg_table}
\alias{cjpg_table}
\title{Downloads information about some CJPG structures}
\usage{
cjpg_table(type, tj = "tjsp")
}
\arguments{
\item{type}{Either \code{"classes"} or \code{"subjects"} or \code{"courts"}}

\item{tj}{TJ form which to get data (only works with TJSP for now)}
}
\description{
Downloads a table with information about lawsuit
classes, subjects or courts to help with \code{\link[=download_cjpg]{download_cjpg()}}
}
