% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsg_table.R
\name{cjsg_table}
\alias{cjsg_table}
\title{Downloads information about some CJSG structures}
\usage{
cjsg_table(type, tj = "tjsp")
}
\arguments{
\item{type}{Either \code{"classes"} or \code{"subjects"} or \code{"courts"}}

\item{tj}{TJ form which to get data (only works with TJSP for now)}
}
\description{
Downloads a table with information about lawsuit
classes, subjects or courts to help with \code{\link[=download_cjsg]{download_cjsg()}}
}
