% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cposg.R
\name{download_2deg_lawsuit}
\alias{download_2deg_lawsuit}
\title{Download 2nd degree lawsuits}
\usage{
download_2deg_lawsuit(processos, tj = "tjsp", path = "data-raw/cposg",
  overwrite = FALSE)
}
\arguments{
\item{processos}{Lawsuit ID (only numbers)}

\item{tj}{TJ of the lawsuits (only works with TJSP for now)}

\item{path}{Path to the directory where the lawsuit should be downloaded}

\item{overwrite}{Whether to write over already existing HTMLs}
}
\description{
Download second degree lawsuits from TJSP.
}
