% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsg-tjsp.R
\name{download_cjsg}
\alias{download_cjsg}
\title{Dowload CJSG search results}
\usage{
download_cjsg(query, file, classes = "", subjects = "", courts = "",
  trial_start = "", trial_end = "", registration_start = "",
  registration_end = "")
}
\arguments{
\item{query}{Character vector wtih search query}

\item{file}{Path to file where to save search restults}

\item{classes}{Character vector with lawsuit class IDs}

\item{subjects}{Character vector with lawsuit subject IDs}

\item{courts}{Character vector with lawsuit court IDs}

\item{trial_start}{Date when the trial started}

\item{trial_end}{Date when the trial ended}

\item{registration_start}{Date when registration started}

\item{registration_end}{Date then registration ended}
}
\description{
Dowload CJSG search results
}
\seealso{
\code{\link[=cjsg_table]{cjsg_table()}}
}
