% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpopg.R
\name{download_lawsuit}
\alias{download_lawsuit}
\title{Download lawsuits filed in Brazilian Tribinais de Justica
(Justice Courts)}
\usage{
download_lawsuit(id, path = ".")
}
\arguments{
\item{id}{A character vector of one or more lawsuit IDs}

\item{path}{Path to the directory where the lawsuit should be saved}
}
\value{
A character vector with the path to the downloaded lawsuit
}
\description{
This function downloads lawsuits as PDFs. Given a lawsuit ID, and
the path to a directory it will collect the lawsuit, and save it to
the provided directory.
}
\section{About lawsuits}{

The lawsuits contemplated by this funtion have to be filed in a
Brazilian Tribunal de Justica (Justice Court). \code{\link[=download_lawsuit]{download_lawsuit()}}
finds the lawsuit in its state's online Sistema de Automacao de Justica
(Justice Automation System), solves the captcha withholding the
information, and collects the PDF.
}

\section{Implemented TJs}{

Unfortunatelly \code{\link[=download_lawsuit]{download_lawsuit()}} doesn't yet work with all 27 TJs in
Brazil. Here are the ones already implemented:
\itemize{
\item TJAM (Amazonas)
\item TJBA (Bahia)
\item TJSC (Santa Catarina)
\item TJSP (Sao Paulo)
}
}

\examples{
\dontrun{
download_lawsuit("02575182220138040001")
download_lawsuit("0303349-44.2014.8.24.0020")
download_lawsuit("0552486-62.2015.8.05.0001")
}

}
