% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects-2inst.R
\name{list_secoes_2inst}
\alias{list_secoes_2inst}
\title{Sections of the second instance}
\usage{
list_secoes_2inst(tj = "tjsp")
}
\arguments{
\item{tj}{TJ of the table (only works with TJSP for now)}
}
\value{
\code{tibble} with columns \code{pai}, \code{secao} and \code{cod}
}
\description{
Downloads list of sections of the second instance (can be used in \code{\link[=cjsg_parms]{cjsg_parms()}})
}
