% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cjpg.R
\name{parse_cjpg}
\alias{parse_cjpg}
\title{Parse lawsuits extracted from CJPG query}
\usage{
parse_cjpg(file, cores = 1)
}
\arguments{
\item{file}{Character vector with the paths to one or more files}

\item{cores}{Number of cores to use when parsing}
}
\value{
A tibble with the columns
\itemize{
\item \code{file} Name of the file
\item \code{id_lawsuit} Number of the lawsuit (doesn't have to be unique)
\item \code{code_lawsuit} Unique code of the lawsuit
\item \code{subject} Subject of the lawsuit
\item \code{class} Class of the subject
\item \code{district} Name of the district
\item \code{date_available} Date when lawsuit was made available (\%d/\%m/\%Y)
\item \code{jurisdiction} Name of the jurisdiction
\item \code{judge} Name of the judge
\item \code{court} Body responsible for the lawsuit
\item \code{summary} Summary of the ruling
}
}
\description{
Parse lawsuits extracted from CJPG query
}
