% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-cjpg-tjsp.R
\name{parse_cjpg}
\alias{parse_cjpg}
\title{Parser do CJPG}
\usage{
parse_cjpg(arqs)
}
\arguments{
\item{arqs}{vetor de arquivos (caminho completo) a serem lidos.}
}
\value{
tibble com as colunas
\itemize{
\item \code{arq} nome do arquivo lido.
\item \code{id} id contido na p??gina lida.
\item \code{cd_acordao} c??digo ??nico do ac??rd??o.
\item \code{n_processo} n??mero do processo (pode repetir).
\item \code{comarca} nome da comarca.
\item \code{data_julgamento} data de julgamento em formato \%d/\%m/\%Y.
\item \code{data_registro} data de registro no sistem em formato \%d/\%m/\%Y.
\item \code{ementa} ementa do ac??rd??o (muitos vazios).
\item \code{orgao_julgador} c??mara julgadora do recurso.
\item \code{outros_numeros} n??meros antigos / complementares.
\item \code{relatora} Nome do relator ou relatora do recurso.
\item \code{classe_assunto} Classe / assunto, separados por " / ".
\item \code{txt_ementa} Texto da ementa sem formata????o.
}
}
\description{
Parser dos arquivos HTML baixados pela fun????o \code{\link{cjpg}}.
}
