% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_parser.R
\name{parse_parts}
\alias{parse_parts}
\title{Parse parts}
\usage{
parse_parts(parser)
}
\arguments{
\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}
}
\description{
Parse parts
}
