% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsg.R
\name{peek_cjsg}
\alias{peek_cjsg}
\title{Check how long a call to \code{\link[=download_cjsg]{download_cjsg()}} will probably take}
\usage{
peek_cjsg(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[=download_cjsg]{download_cjsg()}} (
\code{path} will be ignored)}
}
\description{
Check how long a call to \code{\link[=download_cjsg]{download_cjsg()}} will probably take
}
\seealso{
\code{\link[=download_cjpg]{download_cjpg()}}, \code{\link[=cjpg_table]{cjpg_table()}}
}
