% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{run_parser}
\alias{run_parser}
\title{Run a parser}
\usage{
run_parser(file, parser, path = ".", cores = 1)
}
\arguments{
\item{file}{A character vector with the paths to one ore more files}

\item{parser}{A parser returned by \code{\link[=make_parser]{make_parser()}}}

\item{path}{The path to a directory where to save RDSs}

\item{cores}{The number of cores to be used when parsing}
}
\description{
Run a parser
}
